Django-signals_orm-0x04/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
